put the prenotification parser into a class
